prompts/requirements.txt
openai 
google-search-results
jupyterlab
pyarrow==10.0.1
snowflake-connector-python
